Separate render fps from logic fps
The renderer should always be fixed to the monitor refresh rate. This
should happen by default with requestAnimationFrame. We now allow
reducing the logic fps below the monitor refresh rate. The main
intention of this is to handle higher refresh rates with the same logic
fps for more game consistency.

Probably could use some work in the future as updates are queue
immediately if catch-up needs to be performed which doesn't result in
much smoother higher fps gameplay at all beyond 60fps.